+++ packages/apps/proof-of-us/src/app/user/page.tsx
@@ -14,8 +14,8 @@ const Page: FC = () => {
       {!isLoading && !error && (
         <ul>
           {data.map((proofOfUs) => (
-            <li key={proofOfUs.id}>
-              <Link href={`/user/proof-of-us/${proofOfUs.id}`}>
+            <li key={proofOfUs.tokenId}>
+              <Link href={`/user/proof-of-us/${proofOfUs.tokenId}`}>
                 {new Date(proofOfUs.date).toLocaleDateString()}
               </Link>
             </li>