+++ packages/libs/client-examples/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kadena/client-examples
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [64dd84ee0]
+- Updated dependencies [bbeef98a6]
+- Updated dependencies [0540b213b]
+- Updated dependencies [3fc8ac86d]
+- Updated dependencies [72f472e58]
+  - @kadena/client-utils@0.3.0
+  - @kadena/client@1.8.0
+
 ## 0.0.10
 
 ### Patch Changes